refactor(code_agent): modularize utils, externalize config, and simplify executor
 Renamed utils.py to processing_utils.py and split it into modular components for easier maintenance  Moved configuration logic from code_agent.py into utils/model_factory.py and utils/settings.py  Simplified executor.py by removing redundant logic and centralizing control All components tested and working correctly at this point in the refactor.
