+++ src/os_assistant/tools/code_agent/wrapper.py
@@ -3,7 +3,7 @@ from langchain_core.tools import tool
 from .core.models import CodeAnalysis
 from .core.run_code import run_code_execution
 from .execution.executors import execute_code_in_subprocess
-from .utils.parsers import ensure_string
+from .processing_utils.string_utils import ensure_string
 
 
 @tool
