+++ src/os_assistant/utils/model_factory.py
@@ -1,5 +1,7 @@
 from langchain_ollama import ChatOllama
+
 import os_assistant.utils.settings as settings
+
 
 def create_model(**overrides):
     return ChatOllama(
@@ -8,8 +10,14 @@ def create_model(**overrides):
         base_url=overrides.get("base_url", settings.MODEL_BASE_URL),
     )
 
+
 # Primary model
 model = create_model()
 
 # Backup model for fixing outputs
 fixing_model = create_model()
+
+# Coding Model
+coding_model = create_model(
+    model=settings.LLM_MODEL_CODING,
+)
